requirements-termux.txt
requests
pycryptodome
python-dotenv
tweepy
github3.py
beautifulsoup4
